legacy/requirements.txt
fastapi
uvicorn
youtube-transcript-api
requests
python-dotenv
markdown
weasyprint
openai
